pdf-extractor/requirements.txt
PyPDF2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.0
